Fix memory bug from multiply allocation
The configure and configureModelDiscretization functions are called
multiple times, which caused multiple allocations for some members of
the unit-operations. This lead to memory leaks, which was especially
severe for the DG code (e.g. 2.5 MB for a single test simulation with
moderate/low discretization) since more members are allocated. This is
fixed by checking if the unit-operation members are already allocated.