Fix AD directions in tests
For the tests, some units need to initialize the AD directions with
zero. This is not done sometimes when multiple tests are run: The number
of AD directions used might still be set from the previous test, leading
to uninitialized value errors.
Thus, we set the AD directions before configuring the units.

Co-authored-by: Jan Breuer <jan.breuer@fz-juelich.de>