Fix multiple bindings configuration bug in binding mode specification (#329)
This commit fixes the configuration of the binding mode(s) when a vector
is being used to specify multiple bindings, e.g. for multiple particle
types. Specifically, when a vector of length >1 for the binding mode (field IS_KINETIC) was specified, the opposite binding modes were actually configured, which is hereby fixed. The configuration for other cases i.e. single value specification was correct.